// src/app/type.ts
export interface Product {
  id: number
  slug: string
  nameBn: string
  category: string
  categoryNameBn: string
  categoryIcon: string
  unit: string
  image: string
  today: number
  yesterday: number
  lastWeek: number
  lastMonth: number
  change: {
    dir: 'up' | 'down' | 'same'
    pct: number
  }
  markets: {
    market: string
    division: string
    min: number
    max: number
  }[]
}